+++ at_secondary/at_persistence_secondary_server/lib/src/notification/at_notification_manager.dart
@@ -11,4 +11,7 @@ abstract class NotificationManagerSpec {
 
   /// Returns if notification can be accepted.
   bool isNotificationAccepted();
+
+  /// Removes the notification from the Notification Keystore.
+  Future<void> remove(String notificationId);
 }